+++ R/utililties.R
@@ -1014,12 +1014,30 @@ vars_from_label <- function(label) {
 
 #' @rdname transform_fun
 #' @export
+#' @importFrom dplyr mutate across
+#' @importFrom tidyselect any_of
 `transform_fun.smooth_estimates` <- function(object, fun = NULL, ...) {
     ## If fun supplied, use it to transform est and the upper and lower interval
     if (!is.null(fun)) {
         fun <- match.fun(fun)
         object <- mutate(object,
                          across(any_of(c("est", "lower_ci", "upper_ci")),
+                                .fns = fun))
+    }
+
+    object
+}
+
+#' @rdname transform_fun
+#' @export
+#' @importFrom dplyr mutate across
+#' @importFrom tidyselect all_of
+`transform_fun.smooth_samples` <- function(object, fun = NULL, ...) {
+    ## If fun supplied, use it to transform est and the upper and lower interval
+    if (!is.null(fun)) {
+        fun <- match.fun(fun)
+        object <- mutate(object,
+                         across(all_of("value"),
                                 .fns = fun))
     }
 
@@ -1076,6 +1094,7 @@ vars_from_label <- function(label) {
 #' @rdname transform_fun
 #' @export
 #' @importFrom dplyr mutate across
+#' @importFrom tidyselect all_of
 `transform_fun.tbl_df` <- function(object, fun = NULL, column = NULL, ...) {
     if (is.null(column)) {
         stop("'column' to modify must be supplied.")